+++ xarray/core/dataset.py
@@ -1910,6 +1910,97 @@ class Dataset(Mapping, ImplementsDatasetReduce, BaseDataObject,
             obj[name] = (dims, data)
         return obj
 
+    def to_dict(self):
+        """
+        Convert this dataset to a dictionary following xarray naming
+        conventions.
+
+        Useful for coverting to json.
+        """
+        d = {'coords': {}, 'attrs': dict(self.attrs), 'dims': dict(self.dims),
+             'data_vars': {}}
+
+        def time_check(val):
+            # needed because of numpy bug GH#7619
+            if np.issubdtype(val.dtype, np.datetime64):
+                val = val.astype('datetime64[us]')
+            elif np.issubdtype(val.dtype, np.timedelta64):
+                val = val.astype('timedelta64[us]')
+            return val
+
+        for k in self.coords:
+            data = time_check(self[k].values).tolist()
+            d['coords'].update({k: {'data': data,
+                                    'dims': self[k].dims,
+                                    'attrs': dict(self[k].attrs)}})
+        for k in self.data_vars:
+            data = time_check(self[k].values).tolist()
+            d['data_vars'].update({k: {'data': data,
+                                       'dims': self[k].dims,
+                                       'attrs': dict(self[k].attrs)}})
+        return d
+
+    @classmethod
+    def from_dict(cls, d):
+        """
+        Convert a dictionary into an xarray.Dataset.
+
+        Input dict can take several forms::
+
+            d = {'t': {'dims': ('t'), 'data': t},
+                 'a': {'dims': ('t'), 'data': x},
+                 'b': {'dims': ('t'), 'data': y}}
+
+            d = {'coords': {'t': {'dims': 't', 'data': t,
+                                  'attrs': {'units':'s'}}},
+                 'attrs': {'title': 'air temperature'},
+                 'dims': 't',
+                 'data_vars': {'a': {'dims': 't', 'data': x, },
+                               'b': {'dims': 't', 'data': y}}}
+
+        where 't' is the name of the dimesion, 'a' and 'b' are names of data
+        variables and t, x, and y are lists, numpy.arrays or pandas objects.
+
+        Parameters
+        ----------
+        d : dict, with a minimum structure of {'var_0': {'dims': [..], \
+                                                         'data': [..]}, \
+                                               ...}
+
+        Returns
+        -------
+        obj : xarray.Dataset
+
+        See also
+        --------
+        xarray.DataArray.from_dict
+        """
+
+        if not set(['coords', 'data_vars']).issubset(set(d)):
+            variables = d.items()
+        else:
+            import itertools
+            variables = itertools.chain(d.get('coords', {}).items(),
+                                        d.get('data_vars', {}).items())
+        try:
+            variable_dict = OrderedDict([(k, (v['dims'],
+                                              v['data'],
+                                              v.get('attrs'))) for
+                                         k, v in variables])
+        except KeyError as e:
+            raise ValueError(
+                "cannot convert dict without the key "
+                "'{dims_data}'".format(dims_data=str(e.args[0])))
+        obj = cls(variable_dict)
+
+        # what if coords aren't dims?
+        coords = set(d.get('coords', {})) - set(d.get('dims', {}))
+        obj = obj.set_coords(coords)
+
+        obj.attrs.update(d.get('attrs', {}))
+
+        return obj
+
     @staticmethod
     def _unary_op(f, keep_attrs=False):
         @functools.wraps(f)