+++ doc/whats-new.rst
@@ -20,9 +20,12 @@ v0.8.2 (unreleased)
 
 Enhancements
 ~~~~~~~~~~~~
-
 - New documentation on :ref:`panel transition`. By
   `Maximilian Roos <https://github.com/MaximilianR>`_.
+- New ``Dataset`` and ``DataArray`` methods :py:meth:`to_dict` and
+  :py:meth:`from_dict` to allow easy conversion between dictionaries and
+  xarray objects (:issue:`432`). See :ref:`dictionary IO <dictionary io>`
+  for more details. By `Julia Signell <https://github.com/jsignell>`_.
 
 Bug fixes
 ~~~~~~~~~